tgbot/requirements.txt
celery
Django<3
psycopg2-binary
python-telegram-bot
redis
requests
django-mptt
uwsgi